environment /language_translation_env.py
import gymnasium as gym
import numpy as np

class LanguageTranslationEnv(gym.Env):
    metadata = {"render_modes": ["human", "rgb_array"]}

    def __init__(self, render_mode=None):
        super(LanguageTranslationEnv, self).__init__()

        # Define action and observation space (update action space to match model)
        self.action_space = gym.spaces.Discrete(5)  # Updated to 5 actions
        self.observation_space = gym.spaces.Box(low=0, high=1, shape=(10,), dtype=np.float32)  # Keep the shape=(10,)

        self.render_mode = render_mode  # Store render mode

    def step(self, action):
        # Implement step logic
        observation = np.random.random(self.observation_space.shape)
        reward = 1.0  # Example reward
        done = False
        info = {}
        truncated = False  # Return False for truncated if no time limit is set

        return observation, reward, done, truncated, info  # Updated to return 5 values

    def reset(self, seed=None, options=None):
        super().reset(seed=seed)
        observation = np.random.random(self.observation_space.shape)
        return observation, {}

    def render(self):
        if self.render_mode == "human":
            print("Rendering environment in human mode")
        elif self.render_mode == "rgb_array":
            return np.zeros((100, 100, 3), dtype=np.uint8)  # Dummy RGB array
