environment /requirements.txt
gym
numpy
PyOpenGL
stable-baselines3
